+++ opendj-sdk/opendj-server-legacy/src/main/java/org/opends/server/backends/jeb/IndexInputBuffer.java
@@ -116,7 +116,7 @@ final class IndexInputBuffer implements Comparable<IndexInputBuffer>
     if (leftToRead < cache.remaining())
     {
       cache.limit((int) (cache.position() + leftToRead));
-      bytesToRead = (int) leftToRead;
+      bytesToRead = leftToRead;
     }
     else
     {
